requirements.txt
tkinter>=8.6
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
cryptography>=3.4.0
pyjwt>=2.4.0
bcrypt>=3.2.0
jsonschema>=4.0.0